% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.snnR.R
\name{print.snnR}
\alias{print.snnR}
\title{Print an snnR object}
\usage{
\method{print}{snnR}(x, ...)
}
\arguments{
\item{x}{An \code{snnR} object from one of the \code{snnR} functions.}

\item{\dots}{Additional arguments; none are used in this method.}
}
\description{
Displays the results a \code{snnR} object.
}

