% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.snnR.R
\name{print.snnR_extended}
\alias{print.snnR_extended}
\title{Print an snnR_extended object}
\usage{
\method{print}{snnR_extended}(x, ...)
}
\arguments{
\item{x}{An \code{snnR_extended} object from one of the \code{snnR_extended} functions.}

\item{\dots}{Additional arguments; none are used in this method.}
}
\description{
Displays the results a \code{snnR_extended} object.
}

